requirements.txt
fastapi~=0.115.12
uvicorn~=0.24.0.post1
python-multipart==0.0.6
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1
pandas==2.2.0
pydantic==2.5.2
jinja2==3.1.2
numpy~=2.2.5
spacy~=3.8.5